requirements.txt
streamlit
tensorflow
numpy
pillow
matplotlib
opencv-python